test: Silence warnings on Win64 builds
(Merged from https://github.com/openssl/openssl/pull/27806)

+++ test/asynciotest.c
@@ -125,7 +125,7 @@ static int async_write(BIO *bio, const char *in, int inl)
 {
     struct async_ctrs *ctrs;
     int ret = 0;
-    size_t written = 0;
+    int written = 0;
     BIO *next = BIO_next(bio);
 
     if (inl <= 0)
@@ -238,7 +238,7 @@ static int async_write(BIO *bio, const char *in, int inl)
         }
         /* Write any data we have left after fragmenting */
         ret = 0;
-        if ((int)written < inl) {
+        if (written < inl) {
             ret = BIO_write(next, in + written, inl - written);
         }
 
@@ -281,7 +281,7 @@ static int async_gets(BIO *bio, char *buf, int size)
 
 static int async_puts(BIO *bio, const char *str)
 {
-    return async_write(bio, str, strlen(str));
+    return async_write(bio, str, (int)strlen(str));
 }
 
 #define MAX_ATTEMPTS    100
