+++ test/ec_internal_test.c
@@ -559,13 +559,13 @@ int setup_tests(void)
     ADD_TEST(ec2m_field_sanity);
     ADD_TEST(field_tests_ec2_simple);
 #endif
-    ADD_ALL_TESTS(field_tests_default, crv_len);
+    ADD_ALL_TESTS(field_tests_default, (int)crv_len);
 #ifndef OPENSSL_NO_EC_NISTP_64_GCC_128
     ADD_TEST(underflow_test);
 #endif
     ADD_TEST(set_private_key);
     ADD_TEST(decoded_flag_test);
-    ADD_ALL_TESTS(ecpkparams_i2d2i_test, crv_len);
+    ADD_ALL_TESTS(ecpkparams_i2d2i_test, (int)crv_len);
     ADD_TEST(named_group_creation_test);
 
     return 1;
